requirements.txt
gitdb==0.6.4
GitPython==2.0.8
requests==2.11.1
smmap==0.9.0
